@@ -7,6 +7,8 @@ import { AuthProvider } from './auth/authContext'
 import { queryClient } from './services/queryClient'
 import { QueryClientProvider } from 'react-query'
 import { ThemeProvider, createTheme } from '@mui/material/styles';
+import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 
 const { palette } = createTheme();
 const { augmentColor } = palette;
@@ -14,8 +16,23 @@ const createColor = (mainColor: any) => augmentColor({ color: { main: mainColor 
 const theme = createTheme({
   palette: {
     secondary: createColor("#1d4d6c"),
-    mode: 'light',
+    // mode: 'light',
+    mode: 'dark',
   },
+  components: {
+    MuiButton: {
+      styleOverrides: {
+        root: {
+          color: '#ffffffe8',
+          border: '1px solid #ffffff80',
+          "&:hover": {
+            backgroundColor: 'rgba(255, 255, 255, 0.08)',
+            border: '1px solid #FFF',
+          }
+        }
+      }
+    }
+  }
 });
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
@@ -23,7 +40,9 @@ ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
     <AuthProvider>
       <QueryClientProvider client={queryClient}>
         <ThemeProvider theme={theme}>
-          <RouterProvider router={router} />
+          <LocalizationProvider dateAdapter={AdapterDayjs}>
+            <RouterProvider router={router} />
+          </LocalizationProvider>
         </ThemeProvider>
       </QueryClientProvider>
     </AuthProvider>
