@@ -14,8 +14,8 @@ import java.io.IOException;
 
 public class AutenticacaoViaTokenFilter extends OncePerRequestFilter {
 	
-	private TokenService tokenService;
-	private UsuarioRepository repository;
+	private final TokenService tokenService;
+	private final UsuarioRepository repository;
 
 	public AutenticacaoViaTokenFilter(TokenService tokenService, UsuarioRepository repository) {
 		this.tokenService = tokenService;
@@ -48,7 +48,7 @@ public class AutenticacaoViaTokenFilter extends OncePerRequestFilter {
 			return null;
 		}
 		
-		return token.substring(7, token.length());
+		return token.substring(7);
 	}
 
 }
