@@ -1,7 +1,7 @@
 package br.com.app.controller.dto;
 
 public class ErrorDto {
-    private String message;
+    private final String message;
 
     public ErrorDto(String message) {
         this.message = message;
