@@ -1,5 +1,5 @@
 package br.com.app.modelo;
 
 public enum Etapa {
-    CRIADO, EM_PROGRESSO, CONCLUIDO
+    PENDENTE_ATIVACAO, CONCLUIDO
 }
